pyproject.toml
[build-system]
build-backend = "poetry.masonry.api"
requires = ["poetry>=0.12"]

[tool]
[tool.poetry]
authors = ["Your Name <you@example.com>"]
description = ""
name = "aquasingledebugmonitor"
version = "0.1.0"
[tool.poetry.dependencies]
Pillow = "^7.0"
flask = "^1.1"
python = "^3.8"
werkzeug = "^0.16.0"
[tool.poetry.dev-dependencies]
